# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a demo version of Gin app.",
        "title": "Gin Demo App",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/garage/add": {
            "post": {
                "description": "Add a new garage to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Garage"
                ],
                "summary": "Add a new garage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Garage request payload",
                        "name": "garageReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddGarageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Garage added successful",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
        "/admin/garage/delete": {
            "delete": {
                "description": "Delete a garage by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Garage"
                ],
                "summary": "Delete a garage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Garage request payload",
                        "name": "garageReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeletGarageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Garage deleted successful",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "404": {
                        "description": "Garage not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
        "/admin/garage/update": {
            "put": {
                "description": "Update a garage by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Garage"
                ],
                "summary": "Update a garage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update request payload",
                        "name": "updateReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateGarageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Garage updated successful",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "404": {
                        "description": "Garage not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
...
        "/garage/rarity": {
            "get": {
                "description": "Retrieve the list of rarities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Garage"
                ],
                "summary": "Get All Rarity list",
                "responses": {
                    "200": {
                        "description": "List fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
        "/garage/types": {
            "get": {
                "description": "Retrieve the list of all garages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Garage"
                ],
                "summary": "Get All Garage type List",
                "responses": {
                    "200": {
                        "description": "Garage type list fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
        "/garages/get-all": {
            "get": {
                "description": "Retrieve the list of all garages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Garage"
                ],
                "summary": "Get All Garage List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records to skip (default is 0)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of records to fetch (default is 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Garage list fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.AddGarageRequest": {
            "type": "object",
            "properties": {
                "coinsRequired": {
                    "description": "coins required to unlock the garage",
                    "type": "integer"
                },
                "garageName": {
                    "type": "string"
                },
                "garageType": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "level": {
                    "description": "level required to unlock the garage",
                    "type": "integer"
                },
                "longitude": {
                    "type": "number"
                },
                "rarity": {
                    "type": "integer"
                }
            }
        },
...
        "request.DeletGarageReq": {
            "type": "object",
            "properties": {
                "garageId": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateGarageReq": {
            "type": "object",
            "properties": {
                "coinsRequired": {
                    "description": "coins required to unlock the garage",
                    "type": "integer"
                },
                "garageId": {
                    "type": "string"
                },
                "garageName": {
                    "type": "string"
                },
                "garageType": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "level": {
                    "description": "level required to unlock the garage",
                    "type": "integer"
                },
                "longitude": {
                    "type": "number"
                },
                "rarity": {
                    "type": "integer"
                }
            }
        },
...
        "response.Success": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}